requirements.txt
Flask==2.2.5
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
lightgbm==4.3.0
tensorflow==2.15.0
keras-tcn==3.5.0
joblib==1.3.2


